backend/requirements.txt
flask==2.3.0
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
werkzeug==2.3.0
google-generativeai==0.3.0